# ARROW-10851: [C++] Reduce size of generated code for sort kernels