Reduce generated code size and compilation time for `vector_sort.cc` by 35%.

Before:
```
$ ls -la ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
-rw-rw-r-- 1 antoine antoine 29292160 déc.   9 10:12 ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
$ size ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
   text	   data	    bss	    dec	    hex	filename
1191948	   2864	    360	1195172	 123ca4	./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
$ rm ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o && time CCACHE_DISABLE=1 ninja
[65/65] Linking CXX executable release/arrow-flight-benchmark

real	0m40,730s
user	0m44,445s
sys	0m2,286s
```

After:
```
$ ls -la ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
-rw-rw-r-- 1 antoine antoine 19131928 déc.   9 10:15 ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
$ size ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
   text	   data	    bss	    dec	    hex	filename
 763009	   2528	    360	 765897	  bafc9	./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o
$ rm ./src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/vector_sort.cc.o && time CCACHE_DISABLE=1 ninja
[65/65] Linking CXX executable release/arrow-flight-benchmark

real	0m25,740s
user	0m29,434s
sys	0m2,209s
```