@wesm 2020-12-15
Aside: have we checked in on wheel sizes or other bottom line code size metrics lately?