@pitrou 2020-12-15
Apparently we're at 18.1MB for the latest nightly build: https://dev.azure.com/ursa-labs/crossbow/_build/results?buildId=21639&view=logs&j=0da5d1d9-276d-5173-c4c4-9d4d4ed14fdb&t=b525c197-f769-5e52-d38a-e6301f5260f2

This doesn't change much from 2.0.0 (17.8MB).